dto 수정

+++ ReciPT/src/main/java/samdasu/recipt/controller/dto/Review/GptReviewResponseDto.java
@@ -22,7 +22,7 @@ public class GptReviewResponseDto {
     private String title;
     @NotNull
     private String comment;
-    private Integer viewCount;
+    private Long viewCount;
     private Integer likeCount;
     private GptRecipe gptRecipe;
     private List<ImageFileResponseDto> imageFiles;
@@ -40,7 +40,7 @@ public class GptReviewResponseDto {
                 .collect(Collectors.toList());
     }
 
-    public GptReviewResponseDto(Long reviewId, String username, String title, String comment, Integer viewCount, Integer likeCount, GptRecipe gptRecipe, List<ImageFileResponseDto> imageFiles) {
+    public GptReviewResponseDto(Long reviewId, String username, String title, String comment, Long viewCount, Integer likeCount, GptRecipe gptRecipe, List<ImageFileResponseDto> imageFiles) {
         this.reviewId = reviewId;
         this.username = username;
         this.title = title;
@@ -51,7 +51,7 @@ public class GptReviewResponseDto {
         this.imageFiles = imageFiles;
     }
 
-    public static GptReviewResponseDto createGptReviewResponseDto(Long reviewId, String username, String title, String comment, Integer viewCount, Integer likeCount, GptRecipe gptRecipe, List<ImageFileResponseDto> imageFiles) {
+    public static GptReviewResponseDto createGptReviewResponseDto(Long reviewId, String username, String title, String comment, Long viewCount, Integer likeCount, GptRecipe gptRecipe, List<ImageFileResponseDto> imageFiles) {
         return new GptReviewResponseDto(reviewId, username, title, comment, viewCount, likeCount, gptRecipe, imageFiles);
     }
